+++ aws_lambda_powertools/utilities/parser/envelopes/base.py
