+++ aws_lambda_powertools/utilities/parser/parser.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, Dict, Optional
 from pydantic import BaseModel, ValidationError
 
 from aws_lambda_powertools.middleware_factory import lambda_handler_decorator
-from aws_lambda_powertools.utilities.advanced_parser.envelopes import Envelope, parse_envelope
+from aws_lambda_powertools.utilities.parser.envelopes import Envelope, parse_envelope
 
 logger = logging.getLogger(__name__)
 
@@ -21,11 +21,11 @@ def parser(
 
         As Lambda follows (event, context) signature we can remove some of the boilerplate
         and also capture any exception any Lambda function throws as metadata.
-        event will be the parsed and passed as a BaseModel pydantic class of the input type "schema" 
+        event will be the parsed and passed as a BaseModel pydantic class of the input type "schema"
         to the lambda handler.
         event will be extracted from the envelope in case envelope is not None.
         In case envelope is None, the complete event is parsed to match the schema parameter BaseModel definition.
-        In case envelope is not None, first the event is parsed as the envelope's schema definition, and the user 
+        In case envelope is not None, first the event is parsed as the envelope's schema definition, and the user
         message is extracted and parsed again as the schema parameter's definition.
 
         Example
@@ -41,9 +41,9 @@ def parser(
         handler:  input for lambda_handler_decorator, wraps the handler lambda
         event:    AWS event dictionary
         context:  AWS lambda context
-        schema:   pydantic BaseModel class. This is the user data schema that will replace the event. 
+        schema:   pydantic BaseModel class. This is the user data schema that will replace the event.
                   event parameter will be parsed and a new schema object will be created from it.
-        envelope: what envelope to extract the schema from, can be any AWS service that is currently 
+        envelope: what envelope to extract the schema from, can be any AWS service that is currently
                   supported in the envelopes module. Can be None.
 
         Raises
